Revert "refactor: add computer repository"
